renamed some service files to be in dash case
