+++ src/pages/IntegrationsPage.vue
@@ -102,7 +102,7 @@
   import {useFindPaginate} from '../';
   import SelectOrAdd from '../components/formGeneratorCustom/SelectOrAdd';
   import {computed, ref} from 'vue';
-  import {IntegrationAuths} from '../stores/services/integrationAuths';
+  import {IntegrationAuths} from '../stores/services/integration-auths';
 
   export default {
     name: 'integrations-page',
