+++ src/stores/index.js
@@ -6,14 +6,14 @@ import domains, { Domains } from './services/domains';
 import environments, { Environments } from './services/environments';
 import features, { Features } from './services/features';
 import geocode, { Geocode } from './services/geocode';
-import glClasses, { GLClasses } from './services/glClasses';
-import glDepartments, { GLDepartments } from './services/glDepartments';
+import glClasses, { GLClasses } from './services/gl-classes';
+import glDepartments, { GLDepartments } from './services/gl-departments';
 import hosts, { Hosts } from './services/hosts';
 import instances, { Instances } from './services/instances';
-import integrationAuths, { IntegrationAuths } from './services/integrationAuths';
+import integrationAuths, { IntegrationAuths } from './services/integration-auths';
 import integrations, { Integrations } from './services/integrations';
 import placesAutoComplete, { PlacesAutoComplete } from './services/places-auto-complete';
-import quickbookCompanies, { QuickbooksCompanies } from './services/quickbookCompanies';
+import quickbookCompanies, { QuickbooksCompanies } from './services/quickbook-companies';
 import vInstances, { VInstances } from './services/vInstances';
 import wallets, { Wallets } from './services/wallets';
